+++ Eset/CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## 2026-07-03 - 1.1.0
+
+### Added
+
+- Add ESET vulnerability asset connector.
+- Add ESET user asset connector.
+
 ## 2025-06-28 - 1.0.2
 
 ### Added